let c-space stop/start drawing on the framebuffer
Using ^L to redraw the framebuffer was really annoying because most
programs bind this key.  What's more, fbvnc used to update the screen
when fbpad was viewing another tag.  After this change c-space disables
updating the screen and when pressed again enables it and redraws the
screen.  This patch combined with the scrsnap branch of fbpad on which
fbvnc can be run in a saved tag, is very nicer to use.